+++ src/slices.rs
@@ -2,31 +2,34 @@
 !*/
 
 use super::Endian;
-use std::ptr;
 
 /// Traverse a slice, performing the `Endian` method on each item in place.
 impl<'a, T: Endian> Endian for &'a mut [T] {
 	fn from_be(self) -> Self {
 		for elt in self.iter_mut() { unsafe {
-			ptr::write(elt, ptr::read(elt).from_be());
+			let ptr = elt as *mut T;
+			ptr.write(ptr.read().from_be());
 		} }
 		self
 	}
 	fn from_le(self) -> Self {
 		for elt in self.iter_mut() { unsafe {
-			ptr::write(elt, ptr::read(elt).from_le());
+			let ptr = elt as *mut T;
+			ptr.write(ptr.read().from_le());
 		} }
 		self
 	}
 	fn to_be(self) -> Self {
 		for elt in self.iter_mut() { unsafe {
-			ptr::write(elt, ptr::read(elt).to_be());
+			let ptr = elt as *mut T;
+			ptr.write(ptr.read().to_be());
 		} }
 		self
 	}
 	fn to_le(self) -> Self {
 		for elt in self.iter_mut() { unsafe {
-			ptr::write(elt, ptr::read(elt).to_le());
+			let ptr = elt as *mut T;
+			ptr.write(ptr.read().to_le());
 		} }
 		self
 	}