+++ tests/complex_struct.rs
@@ -5,7 +5,6 @@ merely UB to transmute them into byte arrays and access into them like that,
 which is what all the other tests are doing.
 !*/
 
-#[macro_use]
 extern crate endian_trait;
 
 use endian_trait::Endian;