+++ tests/generics.rs
@@ -14,7 +14,6 @@ struct ExampleGeneric<A> where A: Endian {
 So long as this bound is satisfied, Endian can be derived on any generic struct.
 !*/
 
-#[macro_use]
 extern crate endian_trait;
 
 use endian_trait::Endian;